Extract `QuestionDashboardPicker` component
